move some content into other files
